Allocate config args from new rather than malloc
ASan reported a alloc-dealloc-mismatch in Config::load_args because
memory was allocated with malloc and stored in unique_ptr which later
freed it with delete. new should be paired with delete and malloc should
be paired with free. This addresses this by allocating the arg buffer
with new rather than malloc.